Renaming files to conform to naming standard
